(Hopefully) completed Grades activity - each subject view comes with a delete subject and add grade button. Delete subjects deletes the subject, add grade shows a popup which lets the user select a grade from 1 - 5 and add it to the subject whose add grade button activated the popup.

+++ app/src/main/java/com/mrmi/schoolhelper/CustomSubject.java
@@ -1,10 +1,7 @@
 package com.mrmi.schoolhelper;
 
-import android.view.View;
 import com.google.gson.annotations.SerializedName;
-
 import java.util.ArrayList;
-import java.util.List;
 
 public class CustomSubject
 {
@@ -15,10 +12,10 @@ public class CustomSubject
     public int averageGrade;
 
     @SerializedName("Grades")
-    public ArrayList<Integer> gradeList;
+    public ArrayList<Short> gradeList;
 
     //Constructor
-    public CustomSubject(String name, int average, ArrayList<Integer> grades)
+    public CustomSubject(String name, int average, ArrayList<Short> grades)
     {
         subjectName=name;
         averageGrade=average;
